E2E logcheck: move to "invariants" directory
This check needs closer scrutiny because it hooks into the overall test suite execution.
